[Feat] 공지 상세 화면 라우팅 연결, UI 구현
