+++ yappu-world-ios/Source/Presentation/Home/Router/HomeNavigationView.swift
@@ -17,23 +17,23 @@ struct HomeNavigationView: View {
     
     var body: some View {
         NavigationStack(path: $router.path) {
-            if let homeViewModel = router.homeViewModel {
-                HomeView(viewModel: homeViewModel)
-                    .navigationDestination(for: HomePath.self) { path in
-                        switch path {
-                        case .setting:
-                            if let viewModel = router.settingViewModel {
-                                SettingView(viewModel: viewModel)
-                            }
-                        case .noticeList:
-                            if let viewModel = router.noticeViewModel {
-                                NoticeView(viewModel: viewModel)
-                            }
-                        case .noticeDetail:
-                            NoticeDetailView()
+            HomeView(viewModel: router.homeViewModel)
+                .navigationDestination(for: HomePath.self) { path in
+                    switch path {
+                    case .setting:
+                        if let viewModel = router.settingViewModel {
+                            SettingView(viewModel: viewModel)
+                        }
+                    case .noticeList:
+                        if let viewModel = router.noticeViewModel {
+                            NoticeView(viewModel: viewModel)
+                        }
+                    case .noticeDetail:
+                        if let viewModel = router.noticeDetailViewModel {
+                            NoticeDetailView(viewModel: viewModel)
                         }
                     }
-            }
+                }
         }
         .task {
             await router.onAppear()
