+++ yappu-world-ios/Source/Presentation/Notice/View/NoticeDetailView.swift
@@ -8,17 +8,50 @@
 import SwiftUI
 
 struct NoticeDetailView: View {
+    
+    @State
+    var viewModel: NoticeDetailViewModel
+    
     var body: some View {
-        YPScrollView(axis: .vertical, showsIndicators: true, content: {
+        VStack(alignment: .leading, spacing: 8) {
+            Group {
+                noticeTitleView
+                    .padding(.top, 16)
+                NoticeBadge(notice: viewModel.notice)
+                
+            }
+            .padding(.horizontal, 20)
             
-        })
+            YPScrollView(axis: .vertical, showsIndicators: true, content: {
+                content
+                    .padding(.horizontal, 20)
+            })
+            .ignoresSafeArea(edges: .bottom)
+        }
+        
+    }
+    
+
+}
+
+
+extension NoticeDetailView {
+    var noticeTitleView: some View {
+        VStack {
+            Text(viewModel.notice.title)
+                .font(.pretendard18(.semibold))
+                .foregroundStyle(Color.labelGray)
+            
+        }
+    }
+    
+    var content: some View {
+        Text(viewModel.notice.content)
+            .font(.pretendard15(.regular))
+            .foregroundStyle(.labelGray)
     }
 }
 
-extension NoticeDetailView {
-    
-}
-
 #Preview {
-    NoticeDetailView()
+    NoticeDetailView(viewModel: .init(id: ""))
 }
