+++ yappu-world-ios/Source/Util/NavigationRouter/Home/HomePath.swift
@@ -7,8 +7,8 @@
 
 import Foundation
 
-enum HomePath {
+enum HomePath: Hashable {
     case setting
     case noticeList
-    case noticeDetail
+    case noticeDetail(id: String)
 }
